+++ server/src/weaverbird/pipeline/steps/append.py
@@ -1,11 +1,16 @@
-from typing import Literal
+from typing import TYPE_CHECKING, Literal, Self, TypeVar
 
 from weaverbird.pipeline.steps.utils.base import BaseStep
 from weaverbird.pipeline.steps.utils.render_variables import StepWithVariablesMixin
 
+if TYPE_CHECKING:
+    from weaverbird.pipeline.pipeline import Pipeline, PipelineWithVariables
+
+    PipelineType = TypeVar("PipelineType", bound=Pipeline | PipelineWithVariables)
+
 from .utils.combination import (
-    PipelineOrDomainName,
-    PipelineWithRefsOrDomainNameOrReference,
+    PipelineOrDomainNameOrReference,
+    PipelineWithVariablesOrDomainNameOrReference,
     ReferenceResolver,
     resolve_if_reference,
 )
@@ -16,22 +21,20 @@ class BaseAppendStep(BaseStep):
 
 
 class AppendStep(BaseAppendStep):
-    pipelines: list[PipelineOrDomainName]
+    pipelines: list[PipelineOrDomainNameOrReference]
 
-
-class AppendStepWithVariable(AppendStep, StepWithVariablesMixin):
-    ...
-
-
-class AppendStepWithRefs(BaseAppendStep):
-    pipelines: list[PipelineWithRefsOrDomainNameOrReference]
-
-    async def resolve_references(self, reference_resolver: ReferenceResolver) -> AppendStepWithVariable | None:
+    async def resolve_references(
+        self, reference_resolver: ReferenceResolver, parent_pipeline: "PipelineType"
+    ) -> Self | None:
         resolved_pipelines = [await resolve_if_reference(reference_resolver, p) for p in self.pipelines]
         resolved_pipelines_without_nones = [p for p in resolved_pipelines if p is not None]
         if len(resolved_pipelines_without_nones) == 0:
             return None  # skip the step
-        return AppendStepWithVariable(
+        return self.__class__(
             name=self.name,
             pipelines=resolved_pipelines_without_nones,
         )
+
+
+class AppendStepWithVariable(AppendStep, StepWithVariablesMixin):
+    pipelines: list[PipelineWithVariablesOrDomainNameOrReference]