+++ server/src/weaverbird/pipeline/steps/domain.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Literal, Union
+from typing import TYPE_CHECKING, Literal, TypeVar
 
 from weaverbird.pipeline.steps.utils.base import BaseStep
 from weaverbird.pipeline.steps.utils.combination import (
@@ -8,33 +8,28 @@ from weaverbird.pipeline.steps.utils.combination import (
 )
 
 if TYPE_CHECKING:
-    from weaverbird.pipeline.pipeline import PipelineWithVariables
+    from weaverbird.pipeline.pipeline import Pipeline, PipelineWithVariables
+
+    PipelineType = TypeVar("PipelineType", bound=Pipeline | PipelineWithVariables)
 
 
-class BaseDomainStep(BaseStep):
+class DomainStep(BaseStep):
     name: Literal["domain"] = "domain"
-
-
-class DomainStep(BaseDomainStep):
-    domain: str
-
-
-class DomainStepWithRef(BaseDomainStep):
     domain: str | Reference
 
     async def resolve_references(
-        self, reference_resolver: ReferenceResolver
-    ) -> Union[DomainStep, "PipelineWithVariables"]:
+        self, reference_resolver: ReferenceResolver, parent_pipeline: "PipelineType"
+    ) -> "DomainStep | PipelineType":
         """
         This resolution can return a whole pipeline, which needs to replace the step.
         Not that the resulting array must be flattened:
         it should look like [step 1, step 2, step 3], not [[step 1, step 2], step 3]
         """
-        from weaverbird.pipeline.pipeline import PipelineWithRefs, ReferenceUnresolved
+        from weaverbird.pipeline.pipeline import ReferenceUnresolved
 
         resolved = await resolve_if_reference(reference_resolver, self.domain)
         if isinstance(resolved, list):
-            return await PipelineWithRefs(steps=resolved).resolve_references(reference_resolver)
+            return await parent_pipeline.__class__(steps=resolved).resolve_references(reference_resolver)
         elif resolved is None:
             raise ReferenceUnresolved()
         else: